+++ lib/stimulus_spec/matchers.rb
@@ -7,6 +7,7 @@ require_relative "matchers/have_stimulus_action"
 require_relative "matchers/have_stimulus_target"
 require_relative "matchers/have_stimulus_value"
 require_relative "matchers/have_stimulus_class"
+require_relative "matchers/have_stimulus_outlet"
 
 module StimulusSpec
   module Matchers